\name{mappal_abs}
\alias{mappal_abs}
\title{helper function for \code{mappal}}
\usage{
mappal_abs(x, pal_vec, max_abs_x = max(abs(x)), type = "div")
}
\description{
This is primarily a helper function for \code{mappal}.
Users should directly call \code{mappal} rather than
calling \code{mappal_abs}.
}

