\docType{data}
\name{marioMat.rda}
\alias{marioMat.rda}
\title{Picture of Mario doing an SVD with only his head}
\description{
  A sample image for testing out image0 plotting. The jpeg
  is from (http://aaronjfisher.github.io). At the time of
  this write-up, this image is located under the CV page
  (http://aaronjfisher.github.io/pages/cv.html)
}
\keyword{data}
\keyword{datasets}

