\name{progCode}
\alias{progCode}
\title{Code snippet for making progress bars}
\usage{
  progCode()
}
\value{
  sample code for writing a progress bar
}
\description{
  I always forget the functions for making progress bars in
  loops. This function reminds me!
}
\examples{
progCode()
#type "progCode" to see the code itself
}

